api/requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.0
flask-cors==3.0.10
Flask-PyMongo==2.3.0
pymongo==3.12.3
Flask-Bcrypt==1.0.1
python-dotenv==0.20.0
gunicorn == 20.0.4
click==8.0.3
dnspython==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Werkzeug==2.0.2